+++ backend/src/main/kotlin/org/tormap/config/properties/IpLookupConfig.kt
@@ -1,4 +1,4 @@
-package org.tormap.config
+package org.tormap.config.properties
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.boot.context.properties.ConstructorBinding
